Allows decimal values for size and position in format panel

+++ javascript/examples/grapheditor/www/js/Format.js
@@ -1857,7 +1857,7 @@ ArrangePanel.prototype.addGeometryHandler = function(input, fn)
 	{
 		if (input.value != '')
 		{
-			var value = parseInt(input.value);
+			var value = parseFloat(input.value);
 
 			if (value != initialValue)
 			{
